logstorage: fix memory leak in preparation phase
When prepare a log file for each filter, working_file_name in current
code are updated to the newest_filename without checking they are freed
or not before. And this leads to memory leak due to not freeing.

By checking working_file_name value are NULL or not before updating will
avoid this memory leak.

Signed-off-by: Dinh Cong Toan <Toan.DinhCong@vn.bosch.com>